Added a lot of functionality
Made a pre-process.py file with ~19 methods to prepare the data. On scraper_module re-wrote the basic algorithm to work more efficiently,  in that file added methods to capture date/hours of work and re-name the folder in order to use them. Added a file settings.py for global variables. Added 2 files for database operations. config.py has the credentials to enter the db but on local level(like an .env) and last the database.py which connects to db and has a sql query to return a date